Revert "lib7z update"
This reverts commit 93297a8420ccb03e384ca3cf8e9be041db2621ad.

+++ src/burner/macos/burner_macos.h
@@ -12,7 +12,7 @@ typedef struct tagRECT {
 } RECT,*PRECT,*LPRECT;
 typedef const RECT *LPCRECT;
 
-typedef unsigned int DWORD;
+typedef unsigned long DWORD;
 typedef unsigned char BYTE;
 
 #ifndef MAX_PATH
